requirements.txt
langchain~=0.2.0
langchain-chroma~=0.1.0
langchain-community~=0.2.0
langchain-core~=0.2.0
langchain-openai~=0.1.0
langchain-text-splitters~=0.2.0
PyYAML
